Move irk parsing to nimble_rssi component
