+++ esphome_nimble/components/nimble_rssi/nimble_rssi_sensor.cpp
@@ -8,9 +8,16 @@ namespace esphome
 
         bool NimbleRssiSensor::parse_device(NimBLEAdvertisedDevice *advertised_device)
         {
-            // TODO: implement irk logic here
-            return true;
-        }
+            if (advertised_device->getAddressType() != BLE_ADDR_RANDOM)
+            {
+                return false;
+            }
+
+            auto address = advertised_device->getAddress();
+            auto naddress = address.getNative();
+
+            return ble_ll_resolv_rpa(naddress, this->irk_);
+        };
 
     } // namespace nimble_tracker
