+++ esphome_nimble/components/nimble_rssi/nimble_rssi_sensor.h
@@ -4,6 +4,9 @@
 #include "esphome/components/sensor/sensor.h"
 #include "esphome/components/nimble_tracker/nimble_tracker.h"
 
+#include "string_utils.h"
+#include "irk_utils.h"
+
 namespace esphome
 {
     namespace nimble_rssi
@@ -11,10 +14,15 @@ namespace esphome
         class NimbleRssiSensor : public sensor::Sensor, public Component, public nimble_tracker::NimbleDeviceListener
         {
         public:
-            void set_irk(const char *irk)
+            void set_irk(std::string irk_hex)
             {
                 this->match_by_ = MATCH_BY_IRK;
-                this->irk_ = irk;
+                this->irk_ = new uint8_t[16];
+
+                if (!hextostr(irk_hex.c_str(), this->irk_, 16))
+                {
+                    ESP_LOGE("nimble_rssi", "Something is wrong with the irk!");
+                }
             }
 
             bool parse_device(NimBLEAdvertisedDevice *advertised_device) override;
@@ -25,7 +33,8 @@ namespace esphome
                 MATCH_BY_IRK,
             };
             MatchType match_by_;
-            const char *irk_;
+
+            uint8_t *irk_;
         };
 
     } // namespace nimble_tracker
